+++ CHANGELOG.md
@@ -11,6 +11,7 @@ nav_order: 1
 
 - Add additional disk space support
 - Update disk space on refresh
+- Remove the `retention_hours` logic from the `aiven_kafka_topic` resource
 
 ## [3.5.1] - 2022-08-16